Move type definitions to types module
