+++ src/components/forms/PlanForm/helpers.ts
@@ -75,7 +75,12 @@ import {
   GoalPriorityType,
   GoalUnit,
   PlanActionCodesType,
+  PlanActivityExpression,
+  PlanActivityFormFields,
   PlanActivityTitlesType,
+  PlanActivityTrigger,
+  PlanFormFields,
+  PlanJurisdictionFormFields,
   taskGenerationStatusType,
 } from './types';
 
@@ -130,60 +135,6 @@ export const PlanSchema = Yup.object().shape({
   title: Yup.string().required(REQUIRED),
   version: Yup.string(),
 });
-
-interface PlanActivityExpression {
-  description: string;
-  expression: string;
-}
-
-interface PlanActivityTrigger {
-  description?: string;
-  expression?: string;
-  name: string;
-}
-
-/** Plan activity form fields interface */
-export interface PlanActivityFormFields {
-  actionCode: string;
-  actionDescription: string;
-  actionIdentifier: string;
-  actionReason: string;
-  actionTitle: string;
-  condition?: PlanActivityExpression[];
-  goalDescription: string;
-  goalDue: Date;
-  goalPriority: string;
-  goalValue: number;
-  timingPeriodEnd: Date;
-  timingPeriodStart: Date;
-  trigger?: PlanActivityTrigger[];
-}
-
-/** Plan jurisdictions form fields interface */
-export interface PlanJurisdictionFormFields {
-  id: string;
-  name: string;
-}
-
-/** Plan form fields interface */
-export interface PlanFormFields {
-  activities: PlanActivityFormFields[];
-  caseNum?: string;
-  date: Date;
-  end: Date;
-  fiReason?: FIReasonType;
-  fiStatus?: FIStatusType;
-  identifier: string;
-  interventionType: InterventionType;
-  jurisdictions: PlanJurisdictionFormFields[];
-  name: string;
-  opensrpEventId?: string;
-  start: Date;
-  status: PlanStatus;
-  taskGenerationStatus: taskGenerationStatusType;
-  title: string;
-  version: string;
-}
 
 /**
  * Convert a plan activity object to an object that can be used in the PlanForm
